requirements.txt
flask
firebase-admin==4.3.0
chardet==3.0.4
click==7.1.2
Cython==0.29.21
dash==1.14.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
decorator==4.4.2
nbformat==5.0.7
numpy==1.19.1
pandas==1.1.0
pandas-datareader==0.9.0
plotly==4.9.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
urllib3==1.25.10
Werkzeug==1.0.1
dash-bootstrap-components==0.10.5
eurostatapiclient==0.2.5
dash_daq==0.5.0
pyrebase4==4.4.0
setuptools==49.6.0
gcloud==0.18.3
